http://cs.brown.edu/courses/cs173/2012/book/types.html
Today is such an embarassment for me. I was supposed to be making a DSL in Haskell and yet I do not feel like it.

So at the very least, I decided to place a few blocks with the energy that I have left. With this I'll call it a day.

There is no greater shame than losing one's drive to me.

I grasped for the sky a bit too early, but unlike with the self improvement algorithm I lovingly named Necromutation, I can access the higher level of functional programming even today.

I will do this.

Even if I did the damn thing in Haskell, I would still have to implement the type inference algorithm in order to assist my neural net agents in the future.

So I might as well do it now.

This is the proof of my determination - and my care.