Right now I am going through the book, trying to get the unification algorithm to work. It might be simple, but it is not easy.
I think there was a bug in the book, but after fixing it, it seems there is another bug as well.

I am going to have to track down the reference implementation from the book in order to see where I am wrong.

But implementing type checking on my own from the compiler does seem much less viable to me now.

I'll see how it goes. I'll also take a look at Benjamin Pierce's Types and Programming Languages book in order to get a gauge at the viability of this approach.