I decided to sever my obligations to this library for the time being. The goal of learning type theory and related stuff which is necessary for my goal of getting better at programming is interfering with my goal of finishing the library so I will cut off the later one.
I need a break from this for a while anyway.

I need to enter this with the same mindset that I had when I started this and I do not feel like fluking around with this thing anymore.

I think the actual ML aspects of the library are not more difficult than making a parser combinator library, but the GPU parts are extremely difficult.

I need something more than F# allows me to express.

For the past 1.5 years, I had a love affair with this language - it was a trully a neverending source of discovery and gains in my programming skill.

But now it is just a hindrance that does not allow me to express what I desire.

And while Haskell would be a step up in the sense that it would allow me to finish the Cuda DSL, it too would not allow to write the code as I want. There is not point in finishing the Cuda DSL in Haskell only to have to write everything else at a lower quality than it would be in F#.

I considered it before, but how about I learn soom Greek and some type theory and some macrology and try giving the advanced aspects of Racket - its macro system - a shot?

I'll forget that I am a machine learning reseracher for the next year if need be and think a little like a type theorist.

There is a way of programming beyond what I can do now, and it is heartbreaking that after I have found it that I cannot use it.

I might be able to manifest my values with some macro magic. I already knew that with macros one could emulate impredicative polymorphism that is needed to make full use of the template pattern, but I did not want to let go of what I have.

And indeed there is no need to let go of my values.

Without a doubt I will exceed what I have here.

Regarding the type inference thing I am working on in `Typechecker_v1.fsx`, while I now know where I went wrong in the `replace_extend` function in the sense that I understand that I actually have to do substitution for constraint solving to work...I do not have a good grasp of what the actual thing should be.

So that is all the more reason to study type theory.

As a reward for all my effort for the past nearly two years, I will loosen the shackles on me for a bit a take a break for a few days. And then I will start this new thing.

I won't update here for a while.