"2pm. I've been deep in thought about this for an entire hour.
I think I have most of it down.

I know how to propagate the arguments upwards using a list as an argument (rather than option like I am doing now), I know what to do with TyV, I understand that instead of having the environment have Ty or Expr, it should be able to hold both Expr and TypedExpr and the Expr should not be just Inlinleable, I know the transformations necessary when the evaluator stumples upon the Inlinleable with arguments on the stack (use TyV here).

Also I need understand now that the evaluator needs to be able to return both Expr and TypedExprs because otherwise how am I going to define Inlinleables?

2:05pm. Right now I am trying to figure out how to do this last thing.

Basically I need a way of managing expectations. That is a very parser combinator kind of thing.

Let me do the chores now and I'll think while I am at it."

The above is an excerpt from my journal.

Making the first typechecker function was easy, but making a function to spit out a type, inlined AST is a bit more difficult.

I need to think about it some more.