I'll call it a day with this.
Even though this is not that complicated, it requires quite some thought. It takes me like 1h just to write a single case.

The only challening one is Inlinleable and then I will be done with v2 of the evaluator.

Maybe I really will succeed in doing it this time?