Did some streamlining of the compiler in order to suspend resumable expressions rather than having to evaluate them from scratch because bugs might creep in once I implement the features that I have in mind. I got wins in surprising places.
In the past few days I've been trying to think of ways of doing higher ranked functions and I finally figure out a way to do it.

"9:15am. I've obsessed really hard all night about the code. It is rare that I generate so much flow.

I need to elaborate on what I've written above.

As usual my thoughts were on higher ranked function passed as lambdas and I've concluded based on thinking how I would represent them in C that they are an absurdity. It can't be done. Higher order functions are not a real thing – they are abstract.

9:20am. So, the only possible way to emulate them would be to pass them as data and JIT compile them, or emulate them using macros by wrapping a function in a macro and then passing the thing into the macro.

No other way, but later one seems in any way realistic.

As soon as that idea idea occurred to me, I got a determination to see it through where it leads. I already had that idea before of course, but I could not quite put the pieces together.

9:25am. So, while generative macros could emulate higher ranked functions just fine, they would blast the global methods table with copies of identical functions, which is why opted out of the idea in the first place.

That is where the tags come in.

It occurred to me that even if I put in tags like I have now, I am going to get errors with generative macros because they will push different functions into the same slot on the memoization dictionary.

9:30am. And memoization is the key to this whole thing. It is the key to making higher ranked functions work.

The flaw of the macro approach is that it is not capable of memoizing.

So when passing a method into a method, simply add its tag to the method being passed in. That is it, that is the solution.

I could get rid of ApplyType and simply get the arguments types from the arguments themselves this way. And higher order functions would not be actual functions, they would just get memoized and inlined into the method.

The only argument that I could not get rid of would be the return type, but I could just make it optional. In the recursive case, I could just specify it by hand.

The other TODO is related to memoization as well. I was wrong at 11:50pm yesterday. It is not the let that is the problem. It is the fact that I am doing backtracking in it and then have this…
    | Inline(expr,args) ->
        teval {d with args = InlineArgs (args,d.env) :: d.args} expr
    | ApplyType(body,type_vars) ->
        teval {d with args=ApplyTypeArgs(type_vars,d.type_env) :: d.args} body

Those two will not get redone during the resume.

And doing backtracking in let is a bad idea anyway. I'd be better off just introducing a SuspendedLet and SuspendedIf and return that instead.

It vaguely bothered me that I was doing backtracking, but doing the above simply did not occur to me.

But yeah, this is real programming. Up to now, for the past few months I've bashing my head against the limitation of the language I am working in, but this is the real deal.

I never actually expected that with generative macros as a concept and memoization I could get this far."

Here is an excerpt from my journal that describes my thought process. Maybe, just maybe I am on the right path with the Cuda compiler this time.

More than a year ago, I spent a month going through Alex Aiken's Compilers course, but it turns out the things I learnt back then weren't even that relevant to what I am doing now.

Before the new year, when I was struggling with the library, my first attempt at getting some abstraction from the C++ code fragments that I was using was to make a shitty kind of macro system, but this here is definitely the real deal.