"http://matt.might.net/articles/closure-conversion/"
...

"At any rate, capturing variables I believe is not a problem at all. I'll just add them as implicit arguments to every call and I even know how to propagate the information as well.

Though I will need some additional information. In method instantiation, I will have to keep a set of all the variables in the environment and a parent pointer to the outer environment.

So given a method, I'd just look at all the variables used and if they are in the environment, add them to the implicit argument list. Do the same for the parents. If they are not in the variable environment that either means they've been declared locally or in a higher scoped method and therefore do not need to in the implicit argument list. Rather that a list, make it a set. I just need to calculate all the intersections.

Though having methods in methods means that just like having methods in Inlineables, they will all get different tags and I will have to deal with code bloat...but that is not important.

The way I will implement this, methods will really be just tagged memoized typechecked macros."

...

"| a, b -> RError <| sprintf "Expected both sides to be types and to be equal types.\nGot true: %A\nGot false: %A" a b

This thing with with not being able to return macros through if statements sits especially poorly with me.

Yeah, now I absolutely have what I need to make first class functions, but...I decided to remove tuples and I decided to add explicit sequentialization and I can't return macros through if statement and I can't return methods from methods…all of this would have been unnecessary if I desugared to the CPS style.

I'd even get pattern matching for free. What is not to like here?"

From my journal.

I need to do some more reasearch into the CPS style. It has a significant amount of advantages over what I am doing now that maybe I'd be better off not passing up.

It would simplify the core language markedly, but I do not understand it completely just yet, so I need to do some reading.

The Racket book does have a bunch of stuff on it though admittedly I just skimmed that part when I read it months ago. It is time for a refresher.