Enough. These books are a nearly complete waste of time, just like most of those type system studies.
Well, the key is `most` and `nearly`. The trouble is that the useful pieces I am using that I got from my studies are irreplaceable.

I do not think I'll would have been able to skip studying those type systems, but for these books, my hunch is that I can skip them just fine.

I'll take a short break and resume work on the language.

This thing I am working on is what in the future will be called the Spiral language.