Ok, I had enough for today. I wrote plenty of stuff and most of it is new. I'll finish and then test what I have written tomorrow.
I think that apart from this all, I am missing are recursive blocks and then as features go, the core of the language will be complete.

No doubt, I will have to add a lot of magic to make interoperating with C++ smooth.

Well, I'll deal with it is as I go along. Despite everything I am making progress on this thing - I won't leave off to go train in the mountains for a month again.