My brain is fucked today, and my strong will is not really carrying me forward.
This is something that I should have taken me only 2h, but I did not sleep particularly well today and I am going to have to leave it for tomorrow.

I'll plug away at it for a while longer, but I do not think I will manage it. Let me give it a shot at any rate