Let me call it a day here.
I think I know what my problem is. Stress. And the reason why I am having so much difficulty here is not because bad sleep – the bad sleep is because I am having difficulty programming.

Quite frankly, this code is so thick and purposeful. It is not that propagating free variables is hard; it is just that I cannot handle that task while keeping all the previous tasks in my head. It is simply too heavy a burden for me to carry. I need to step back a little.

It is obviously that what I've written here has serious architectural problems. It is not good code. I can't do it all in a single function and in a single pass even though in theory it should be possible.

I've heard about it before. Tomorrow I will take a step back again and instead of focusing right on the problem at hand, what I will do is read up on `nanopasses` and think more carefully how to split all the various features of my compiler into various parts.