"This is not good.
| VV _ as left_arg, Vars right_arg -> parse bind_typedexpr_only acc (left_arg, VV right_arg)

You know instead of doing all that work in the Vars case what I should have done is pull out the parse function and reused this for the Vars case!

Gaaaahhhh!!!

But yes, this is why I have the feeling that it is time to stop. I definitely feel like I am walking on thin ice here. I need to some time for what I've done today to settle in and after that, tomorrow, I will spend some time going through this code and try to optimize it for clarity.

I should definitely do that before moving on to Method.

The Inlineable and Method cases will be very similar in terms of pattern matching, I need to make sure that I am not stupidly duplicating work like today.

At any rate, I am too tired to continue. Will continue tomorrow."