Found a bug. match exp_and_seq d tr, exp_and_seq d fl with // TODO: These two should have their own sequencers.
I've experienced this in the past. I remember when I was writing those ML algorithms by hand how they would keep creeping in like this. There is no doubt about it – my code quality is amateurish here.

But even though I've been thinking about it for a while, I can't think of even a single way to simplify the code here. It is ridiculous.

One hand, it seems that everything is its place, but on the other, I am absolutely certain that the structure as a whole is too monolithic. But I had this problem before when implementing closure conversion and I solved it by using the nanopass pattern.

I have the same problem again. I thought it was fine to have expansion and sequentialization, but now I do not have that, but expansion, sequentialization and pattern matching and once again I have the feeling of trying to cram too many things into too small of a box.

And if I had sequentialization as a separate pass, I would never have had the error in the If case.

…

I was wrong. I am an idiot. I should have seen this coming.

Unless I commit myself to the Nanopass pattern from the start, I will always be tempted to cram too many toys into a too tiny box.

I am not a good programmer – I can only program good.

Ok, with this in mind, let me take a step back and I will go through the Nanopass paper so the lesson sinks in.

Then I will do this properly, with a single pass for every one feature.