"Hmmm...I am thinking, is it really fine to mix before evaluation matching like in Inlieables and after evaluation matching like in the TyVars case currently in that case.
That is the big problem.

It is not that I am mixing pattern matching with sequencing and evaluation. It is that I am messing with evaluation order by having that extra case that is making my head spin.

If I made Inlineable pre-evaluation matching and Method post-evaluation matching then everything would fit smoothly.

For now let me continue reading the dissertation.

http://andykeep.com/pubs/dissertation.pdf
A Nanopass Framework For Commercial Compiler Development

I already watched the video on the nanopass site."