Basically I can't do it like this. Going from what I have now straight to the code for method is quite a stretch. It is not just one step more, but more like 2-3 steps.
What I am going to do is start a new version 'd' and just convert what I have now so it accumulates the intermediates.

Yeah, the Inlineable does not need the intermediates unlike Method, but it will give me a sense for how to do it for Method. There is need to struggle too much over this.