Done with implementing the cases and adjusting closure conversion.
Even without tests, the new backend is now at 400 LOC.

I should get to work on tests next, but I do not feel like it right now.

...Well, apart from that, surprisingly in the end I did put in pattern matching. So the next part would be actual code generation. After that I will have a Cuda backend for the Spiral library.

I won't have the well typed interface, but I'll make do with this.

Actually, returning to work on the library and hopefully finishing it? How sick is that. I can hardly believe it.

I thought for sure that I was stuck in an endless recurring nightmare of making ineffective Cuda backends.

Ah, what the hell, let me do those tests. I wanted to get this done with 5 month ago.