I did it. Now the compiler has proper value tuples. I got rid of Vars (variable arguments) entirely as I could not figure out how to propagate names for them.
The reason why I tried putting in Vars in the first place though considering that tuples are more sensible is because I am not sure how to tuple are array in memory.

Easy when you programming in just C++, but I am not sure how the compiler will arrange when they are structs and such.

I am probably going to write some kind of wrapper thing for this. I almost have it my heat, but not quite yet.

Let me give it a shot and then I will call it a day. It is ridiculous how long this took me. It took me 2/3rds of the day just to decide that the Vars thing is a bad idea.