"No, I have no idea how to do it.
I think maybe if I started from scratch I could make something, but as it is, the typechecker is already too large to make any radical changes in it. Or rather, it is not that the change could not be made, it is just that the idea is still too raw in my mind and the problem with the typechecker is that I can't experiment with it while keeping all the pieces in one place.

I will have to start from scratch again.

I'll do a typechecker without macros or pattern matching, just something to get a grip on recursive type checking."

The above is an excerpt from my journal.

Today I really did jack and I had a lot of haslle with forgetting my Google account password. I spent a lot of time trying alternatives before just resetting it.

It is a bit unnerving that I am using strong, 16 digit passwords and yet I've been able to change them with just putting in a 6 digit phone number.

Although I did pretty much nothing today, I do not get the feeling I slacked off either. I've been scatterbrained the entire day today.

The inspiration last night came to me with all the subtlety of a Thor's hammer and pretty much smashed me.

Hopefully tomorrow I will have more focus. Right now it is time for finishing up the day rather than doing new things. I'll try to play around for a bit though before calling it a day.

It is things like these why it is taking me forever to finish my projects.