This is where I will start. This particular typechecker will be solely to explore the feasibility of my new idea. The one I made has to do sequencing, macro expansion, hoisting and various other things during the typechecking phase.
And it has to prepare the ground floor for the code generation phase.

In contrast, I won't aim at code generation at all with this.

I do not think that those other things are particularly difficult, but like the password thing today, they interfere with my thinking.

V1 of the new typechecker will be here solely so I can hone my focus, laserlike on it – on that particular problem domain.