I prettified the prototype and named the algorithm Reality.
As simple as it is, this might be one my best work. There is no way around the need for a side channel or a stack; I've thought about it.

Thanks to this exercise, I feel confident about extending the main typechecker so it can  infer the types for recursive and mutually recursive functions.

It is a joyous occasion.