I was too hasty yesterday. It turns out I was not completely right about unification being unnecessary. Also I really did a bad job in the main typechecker in my haste.
But now it is done - I work much better when I am rested. I can kick back and admire my handywork for a bit.

I knew the addition to the main typechecker would be easy, but I am glad I did the prototype. Without it I could not focus on the task.

I also found a bug in atomic_add_op (the function had duplicates.)

Now what is next on the list?