Enough, I need to get away from here.
I am super pissed off right now. Pattern matching is supposed to be my brilliant piece of code and yet this stupid pice of shit is not working correctly. And worst, it is so bound together than I can't easily add MSet.

Well, I can, I just need to prime my brain for this first. I was primed to do the map module today, not fiddle with pattern matching - that was supposed to have been a done deal.