Got the type checker operational, removed const and shared from Ty, added printing for method prefixes and tuple definitions.
With this I can now generate workable Cuda code.

I cannot even describe how much I want to get through this idiotic compiler phase and on to reinforcement learning.

God, I want it, I want it a lot!

Training those tabular agents was actually quite fun and I want more of it, much more.

But in the end, if I want to cause the Singularity, I need to keep in mind that the essence of programming is communication. That means that just bringing up those programming challenge problems and drilling the agents will not work.

I will need a language much like the one here and turn in into an interface, and a constraint for the agents until they learn the true language.

The ground I am covering here is something I am going to have to cover eventually and better now than later.

When the time comes, thanks to this typechecker they will have a goodies like algebraic effect handlers to play with, and the standard HM inference will not cut it.