"I need some time to think about this.
Because I added things like sequentialization directly in the type checker, it would actually be quite difficult to...no actually it is not because of that.

Because I have macros for function everything is more difficult than it needs to be.

I am mixing Exprs and TypedExprs and because of that inside the CubBlockReduce I can't really evaluate for fear of spilling things.

Why couldn't Inlineable be a part of TypedExpr? Because it needs a type and the big reason why I am doing typechecking like this, with sequentialization in it.

…Ahhhh...I put in tags directly into the Method and Inlineable...but what if instead it was their type?

`TagT of int64`?

Then I would not need to have any special checking in If statements and I could completely eliminate the Expr branch.

Because I have so much mutable garbage in the particular typechecking pass, I can't do something simple here like try one kind of argument and if that fails, try another.

In parser combinators, I could resume my computation. Because of the way I've organized it, I get no such benefit here.

Also, I had not put attention to it, but the way code generator works, everything would be sequentialized anyway.

Shit, I am never going to be done with it...but I have to do it.

I wish the idea of using tags in types had occurred to me at the beginning.

And this would be the prime opportunity to convert the typechecker so it uses exceptions."

An excerpt from my journal. I need some polymorphism to finish CubBlockReduce, but because I did a shitty job, I can't do it.

Thankfully here I have a wonderful opportunity which - will no doubt take me a lot of time again - to cut the size of the typechecker in half.

But I am basically near the finish line here in terms of design.

All the elements are here, I just have to refine them into their proper form.