It is impossible that this code is anything other than correct. I do not think I will ever exceed this in terms of elegance and simplicity. Sometimes it is not about the language, but about the coder.
I agonized a lot over the past 3.5 weeks about the right way of writing various parts, and even while doing pattern matching the various complexities of it seemed absolutely absurdly complex. I thought there was a simple elegant way of doing it, but the insight eluded me.

Until now.

Time to bring the rest on the same level.