One thing leads to another and it seems I will have to take another step back before I am done.
I need to do another rewrite.

Today an idea occured to me how to get rid of the hacks in ApplyAsIfMain and do principled flattened tuple destructuring which I need to make a proper API and while trying to implement a new way of binding, I found out that closure conversion is incorrect the way it has been implemented.

I woke up way too early today and have not been able to put much energy into programming despite being at it for quite a while. Even moreso, there are days like today when I simply just spend most of my time thinking.

I think I will call it a break early and have a rest. I will implement this thing correctly tomorrow. My initial idea of adding a new global variable to the context is not that great since the type system would be useless in helping me with refactoring.

As I considered a few times in the past, I will move tracking of bound and used variables to the closure conversion function - now I have concrete reasons to do so.

This is a bit disappointing since I really thought I would be done for good with the typechecker this time, but on the bright side once I finish this tomorrow or the day after, I really will have everything down phat. I won't have to use hacks for interfacing with the outside world and will have proper flattened tuple methods.

That will be a big, final step forward towards finishing this.