Done with the new and improved typechecker and codegen. Even I think the code is high quality now. I've been really bogged down yesterday, but after suffering comes inspiration.
It is remarkable how simple the current iteration of the typechecker is compared to what I had before..

I spent the entire of February reading on about types and the entirety of March writing this new backend with most of the effort going into the typechecker.

Every day, I've worked on it without fail.

Here are the things that are left before I can proceed to work on the library proper:

1) Improved Array representations. Right now there are some inneficiencies on the code gen side in how I've decided to represent them.

2) Tuple map. It is a bit annoying how I will have to index manually into everything with the way it works currently.

3) The last map module - map_redocol_map. Once I finish this one, the new backend will get to feature parity with what I had before.

4) Some trivial stuff like ExternC blocks and Includes in the code genrator.

I'll get on with it.