Added tuple_map and improved pattern matching functions to their limit in terms of clarity. I need to work a bit more before I can get closure conversion to a similar level.
I also fixed a bunch of bugs in While statements and made sure arguments are propagated correctly.

It is late and I had enough for the day.

I am almost done with this, now comes the final push.

No doubt about it, this typechecker was significantly harder than the actual ML library.

Now that I can do this, there is probably nothing out there that is beyond my ability to grasp. I will finish it this time.