I refactored closure conversion just a bit.
I have a solution for how to do more efficient array representation at the type level, but it involves complicated transformations that I am not really comfortable with and do not feel like implementing.

The way bindings now work is extremely simple and elegant and effective, doing what I have in mind would pretty much ruin that.

I've been thinking all night about how to do this to the point of not being able to sleep well and even now while I should be working on the last map module so I can finally finish this thing, I am still thinking about it.

I feel lit is possible to go further than my current idea. The issue stems from the difficulties of representing relationships between types.

And I have an idea now what the missing piece is.

In order to represent relationships, you really need the concept of self. The 'self' is what I was missing in my thinking.

Yeah, this is it. To make these arrays make sense, what I need are classes. For the thing I am trying to do here, tuples are flawed because they lack the concept of self.

Then during binding, I could just replace the Self with the specific variable name.

Well, let me just do the last module, even if I could I am not sure if this would be worth implementing at the current juncture.