I am done with the map_redocol_map module and therefore with the compiler as well for the time being. I still have some loose ends left, but nothing particularly challenging.
I finally did it. Now I really can get back to work on the Spiral library.

The plan right now is to wait a bit for the things to settle in and compose that post for /r/types so I can fish for info whether there is some better way of representing array size variables at the type level.

After that I will start looking into integrating this with the library proper.