"Bringing back in these old files pretty much brought the old conflicts back in full.
I just had to lie down and think on for a while, what exactly do I want with the new library. What should be my guiding design principle. It comes down to two different avenues:

1) Static:

While thinking of how I wanted to do the map modules, I realized I would not be able to just put the arrays in directly, rather there would be a long conversion process to the form Cuda could take. This would really kill the speed of calling the individual map kernels because steps needed to bring it into form.

I agonized about this for a bit, and realized that if I did it the same style as the typechecker, I could eliminate all the boundary checking and convert the AST to a speedier one in one feel swoop.

To be honest, it is kind of annoying that F# cannot go all the way and convert the code into a fully efficient construct perhaps using staging, but it would do.

Now that I am writing this though, it occurs to me that I am not exactly sure how I would memoize weights, but there are ways of going around this.

2) Dynamic:

This is how the past iterations of the library were structures. Simply put, do not worry about boundary checks, leave them in because their performance impact will be negligible.

The same goes for those kernel calls. They are asynchronous so it won't be a big deal at any rate.

Expressive power is what truly matters, so prioritize that. Use dictionaries to memoize everything and forget that map_1_2_1 nonsense. A single function for all the maps is all I need. Remove generics from DM as well as they are making the code less generic.

I've actually already decided to pursue this, but I am still working out the details.

Now that I've put things into context, I've realized that trying to optimize those map call parameters by tracking them at the type level is probably a waste of time, but I've spent really very little amount of time working on array representations compared to figuring out how to deal with tuples and pattern matching so it is fine.

Maybe in the future instead of adding that VG type I'll just remove the complicated type stuff and go to the standard representation...but I really think it is no trouble at all. It is a piece of cake. The same goes for the VG type. I spent 2 whole days thinking about it, thinking that it would be impossible to add, but it seems rather trivial right now.

Though, I doubt there will be much difference between a map call with 5 and 20 parameters.

Literally the whole consideration here is that everything is amortized by the asynchronous calls and the fact that I am doing bulk operations on arrays. These are not hot loops here.

In the future there will be a language that will let me have the best of all worlds: expressivity, type safety and performance, but as distasteful as it is, I'll settle for the first one here for now. I'll leave making the true language for the Singularity era."

The above is what I wrote in my journal. Despite how the last paragraphs ends, the battle is still raging inside of me.

...I am going to have to hoist those map statements goddamit, static or dynamic.

It looks like I will do jack today in terms of programming. When getting ready for major undertakings, Like always, I will have to spend a while building up my motivation by agonizing about what I am going to do.

...

I need to remember what is at stake here and what are my programming values. This library is a dragon, I gathered up my courage and tried to slay it so many times, but every time I've ended up getting lost on the way there. It is so frustrating.

3 weeks.

Unfortunately, the code I've written and the feeling that went into it is lost to me right now, so just like I did with the typechecker, I am going to have to allocate a chunk of time.

3 weeks.

I will finish this library in 3 weeks. And that will be the final proof that I am reading to dive into reinforcement learning.